Refactor link-checking behavior, notify on failure

+++ scripts/check-links.sh
@@ -1,10 +1,11 @@
 #!/bin/bash
 set -o nounset -o errexit -o pipefail
 
-BUILD_TYPE=${1-schedule}
-HTTP_SERVER_PORT=8888
+source ./scripts/common.sh
 
-check_links() {
+CHECK_TYPE=${1-local}
+
+check_links_www() {
     # We exclude some links:
 	#     - Our generated API docs have lots of broken links
 	#     - Our available versions page includes links to private repos
@@ -13,7 +14,7 @@ check_links() {
 	#     - Our Visual Studio Marketplace link for the Azure Pipelines task extension,
 	#       although valid and publicly available, is reported as a broken link.
 	#     - A number of synthetic illustrative links come from our examples/tutorials.
-    ./node_modules/.bin/blc http://localhost:$HTTP_SERVER_PORT --recursive --follow --get \
+    npx blc https://www.pulumi.com --recursive --follow --get \
         --exclude "/docs/reference/pkg" \
         --exclude "/docs/get-started/install/versions" \
         --exclude "https://api.pulumi.com/" \
@@ -44,7 +45,7 @@ check_links() {
         --exclude "http://my-bucket-1234567.s3-website.us-west-2.amazonaws.com"
 }
 
-check_get_pulumi_links() {
+check_links_local() {
     # We only link to get.pulumi.com in /docs/get-started/install/ and /docs/get-started/install/versions
     npx blc http://localhost:$HTTP_SERVER_PORT/docs/get-started/install/versions/ --follow \
         --exclude-internal \
@@ -55,40 +56,25 @@ check_get_pulumi_links() {
         --exclude "https://www*"
 }
 
-retry() {
-    local n=1
-    local max=3
-    local delay=15
-    while true; do
-    "$@" && break || {
-        if [[ $n -lt $max ]]; then
-            ((n++))
-            echo "Command failed. Attempt $n/$max:"
-            sleep $delay;
-        else
-            echo "The command has failed after $n attempts." >&2
-            exit 1
-        fi
-    }
-    done
-}
+if [ $CHECK_TYPE = "local" ]; then
 
-# Start an HTTP server listening at the root of the built website.
-yarn run http-server public --port $HTTP_SERVER_PORT &>/dev/null &
-HTTP_SERVER_PID=$!
+    # Start an HTTP server listening at the root of the built website.
+    HTTP_SERVER_PORT=8888
+    yarn run http-server public --port $HTTP_SERVER_PORT &>/dev/null &
+    HTTP_SERVER_PID=$!
 
-# Run the link checker once the HTTP server is listening.
-while ! nc -z localhost $HTTP_SERVER_PORT; do sleep 0.1; done
+    # Run the link checker once the HTTP server is listening.
+    while ! nc -z localhost $HTTP_SERVER_PORT; do sleep 0.1; done
 
-# Re-run the link checker up to three times, to handle the occasional transient failure.
-if [ $BUILD_TYPE = "pull_request" ]
-then
     echo "Checking only get.pulumi.com links"
-    retry check_get_pulumi_links
-else
-    echo "Checking all links"
-    retry check_links
+    retry check_links_local 3 15
+
+    # Kill the server process.
+    kill $HTTP_SERVER_PID
 fi
 
-# Kill the server process.
-kill $HTTP_SERVER_PID
+if [ $CHECK_TYPE = "www" ]; then
+
+    echo "Checking all links on pulumi.com"
+    retry check_links_www 3 15 || post_to_slack "ops-notifications" "Eek! :scream_cat: There are broken links on pulumi.com. See the GitHub Actions log for details. https://github.com/${GITHUB_REPOSITORY}/runs/${GITHUB_RUN_ID}"
+fi
